+++ gaseio/format_string/castep.py
@@ -6,7 +6,7 @@ format_string
 import os
 
 from collections import OrderedDict
-import atomtools
+import atomtools.unit
 
 from .. import ext_types
 from .. import ext_methods