+++ gaseio/format_string/gaussian.py
@@ -11,8 +11,9 @@ import numpy as np
 import pandas as pd
 
 import chemdata
-import atomtools
-from atomtools import Status
+
+import atomtools.unit
+from atomtools.status import Status
 
 
 from .. import ext_types
@@ -792,7 +793,7 @@ FORMAT_STRING = {
                 'selection' : -1,
                 'key' : 'calc_arrays/command',
                 'type' : str,
-                'process' : lambda data, arrays: data.strip(),
+                'process' : lambda data, arrays: re.sub('\s+', ' ', data.strip()),
             },
         }),
         'primitive_data_function' : lambda data, arrays: process_fchk(data, arrays),
@@ -892,23 +893,27 @@ FORMAT_STRING = {
             },
             'alpha_nao_density_matrix' : {
                 # 'prerequisite' : ['_has_alpha_spin', '_has_beta_spin'],
-                'equation' : lambda arrays: ext_methods.datablock_to_numpy(re.sub(r'\s+', '\n', \
-                                (arrays['nao_density_matrix']).strip())).flatten(),
+                'equation' : lambda arrays: ext_methods.lower_diagnal_order_2_square(\
+                                ext_methods.datablock_to_numpy(re.sub(r'\s+', '\n', \
+                                (arrays['nao_density_matrix']).strip())).flatten()),
             },
             'alpha_nao_fock_matrix' : {
                 # 'prerequisite' : ['_has_alpha_spin', '_has_beta_spin'],
-                'equation' : lambda arrays: ext_methods.datablock_to_numpy(re.sub(r'\s+', '\n', \
-                                (arrays['nao_fock_matrix'].split('BETA  SPIN')[0]).strip())).flatten(),
+                'equation' : lambda arrays: ext_methods.lower_diagnal_order_2_square(\
+                                ext_methods.datablock_to_numpy(re.sub(r'\s+', '\n', \
+                                (arrays['nao_fock_matrix'].split('BETA  SPIN')[0]).strip())).flatten()),
             },
             'beta_nao_density_matrix' : {
                 'prerequisite' : ['_has_alpha_spin', '_has_beta_spin'],
-                'equation' : lambda arrays: ext_methods.datablock_to_numpy(re.sub(r'\s+', '\n', \
-                                (arrays['nao_fock_matrix'].split('BETA  SPIN')[1]).strip())).flatten(),
+                'equation' : lambda arrays: ext_methods.lower_diagnal_order_2_square(\
+                                ext_methods.datablock_to_numpy(re.sub(r'\s+', '\n', \
+                                (arrays['nao_fock_matrix'].split('BETA  SPIN')[1]).strip())).flatten()),
             },
             'beta_nao_fock_matrix' : {
                 'prerequisite' : ['_has_alpha_spin', '_has_beta_spin'],
-                'equation' : lambda arrays: ext_methods.datablock_to_numpy(re.sub(r'\s+', '\n', \
-                                (arrays['nao_fock_matrix'].split('BETA  SPIN')[2]).strip())).flatten(),
+                'equation' : lambda arrays: ext_methods.lower_diagnal_order_2_square(\
+                                ext_methods.datablock_to_numpy(re.sub(r'\s+', '\n', \
+                                (arrays['nao_fock_matrix'].split('BETA  SPIN')[2]).strip())).flatten()),
                 'delete' : ['_has_alpha_spin', '_has_beta_spin',
                             'nao_fock_matrix', 'nao_density_matrix'],
             },