+++ gaseio/format_string/mol2.py
@@ -4,13 +4,8 @@ format_string
 
 import re
 from collections import OrderedDict
-from io import StringIO
 
-import numpy as np
-import pandas as pd
 
-import chemdata
-import atomtools
 from .. import ext_types
 from .. import ext_methods